+++ keras_cv/models/__init__.py
@@ -104,13 +104,13 @@ from keras_cv.models.efficientnet_v1 import EfficientNetB4
 from keras_cv.models.efficientnet_v1 import EfficientNetB5
 from keras_cv.models.efficientnet_v1 import EfficientNetB6
 from keras_cv.models.efficientnet_v1 import EfficientNetB7
-from keras_cv.models.efficientnet_v2 import EfficientNetV2B0
-from keras_cv.models.efficientnet_v2 import EfficientNetV2B1
-from keras_cv.models.efficientnet_v2 import EfficientNetV2B2
-from keras_cv.models.efficientnet_v2 import EfficientNetV2B3
-from keras_cv.models.efficientnet_v2 import EfficientNetV2L
-from keras_cv.models.efficientnet_v2 import EfficientNetV2M
-from keras_cv.models.efficientnet_v2 import EfficientNetV2S
+from keras_cv.models.backbones.efficientnet_v2.efficientnet_v2_backbone import EfficientNetV2B0
+from keras_cv.models.backbones.efficientnet_v2.efficientnet_v2_backbone import EfficientNetV2B1
+from keras_cv.models.backbones.efficientnet_v2.efficientnet_v2_backbone import EfficientNetV2B2
+from keras_cv.models.backbones.efficientnet_v2.efficientnet_v2_backbone import EfficientNetV2B3
+from keras_cv.models.backbones.efficientnet_v2.efficientnet_v2_backbone import EfficientNetV2L
+from keras_cv.models.backbones.efficientnet_v2.efficientnet_v2_backbone import EfficientNetV2M
+from keras_cv.models.backbones.efficientnet_v2.efficientnet_v2_backbone import EfficientNetV2S
 from keras_cv.models.mlp_mixer import MLPMixerB16
 from keras_cv.models.mlp_mixer import MLPMixerB32
 from keras_cv.models.mlp_mixer import MLPMixerL16
