--- keras_cv/models/efficientnet_v2.py
@@ -1,1076 +1,0 @@
-# Copyright 2022 The KerasCV Authors
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     https://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-
-"""EfficientNet V2 models for KerasCV.
-
-Reference:
-    - [EfficientNetV2: Smaller Models and Faster Training](https://arxiv.org/abs/2104.00298) (ICML 2021)
-    - [Based on the original keras.applications EfficientNetV2](https://github.com/keras-team/keras/blob/master/keras/applications/efficientnet_v2.py)
-"""  # noqa: E501
-
-import copy
-import math
-
-import tensorflow as tf
-from tensorflow import keras
-from tensorflow.keras import layers
-
-from keras_cv.layers import FusedMBConvBlock
-from keras_cv.layers import MBConvBlock
-from keras_cv.models import utils
-from keras_cv.models.weights import parse_weights
-
-DEFAULT_BLOCKS_ARGS = {
-    "efficientnetv2-s": [
-        {
-            "kernel_size": 3,
-            "num_repeat": 2,
-            "input_filters": 24,
-            "output_filters": 24,
-            "expand_ratio": 1,
-            "se_ratio": 0.0,
-            "strides": 1,
-            "conv_type": 1,
-        },
-        {
-            "kernel_size": 3,
-            "num_repeat": 4,
-            "input_filters": 24,
-            "output_filters": 48,
-            "expand_ratio": 4,
-            "se_ratio": 0.0,
-            "strides": 2,
-            "conv_type": 1,
-        },
-        {
-            "conv_type": 1,
-            "expand_ratio": 4,
-            "input_filters": 48,
-            "kernel_size": 3,
-            "num_repeat": 4,
-            "output_filters": 64,
-            "se_ratio": 0,
-            "strides": 2,
-        },
-        {
-            "conv_type": 0,
-            "expand_ratio": 4,
-            "input_filters": 64,
-            "kernel_size": 3,
-            "num_repeat": 6,
-            "output_filters": 128,
-            "se_ratio": 0.25,
-            "strides": 2,
-        },
-        {
-            "conv_type": 0,
-            "expand_ratio": 6,
-            "input_filters": 128,
-            "kernel_size": 3,
-            "num_repeat": 9,
-            "output_filters": 160,
-            "se_ratio": 0.25,
-            "strides": 1,
-        },
-        {
-            "conv_type": 0,
-            "expand_ratio": 6,
-            "input_filters": 160,
-            "kernel_size": 3,
-            "num_repeat": 15,
-            "output_filters": 256,
-            "se_ratio": 0.25,
-            "strides": 2,
-        },
-    ],
-    "efficientnetv2-m": [
-        {
-            "kernel_size": 3,
-            "num_repeat": 3,
-            "input_filters": 24,
-            "output_filters": 24,
-            "expand_ratio": 1,
-            "se_ratio": 0,
-            "strides": 1,
-            "conv_type": 1,
-        },
-        {
-            "kernel_size": 3,
-            "num_repeat": 5,
-            "input_filters": 24,
-            "output_filters": 48,
-            "expand_ratio": 4,
-            "se_ratio": 0,
-            "strides": 2,
-            "conv_type": 1,
-        },
-        {
-            "kernel_size": 3,
-            "num_repeat": 5,
-            "input_filters": 48,
-            "output_filters": 80,
-            "expand_ratio": 4,
-            "se_ratio": 0,
-            "strides": 2,
-            "conv_type": 1,
-        },
-        {
-            "kernel_size": 3,
-            "num_repeat": 7,
-            "input_filters": 80,
-            "output_filters": 160,
-            "expand_ratio": 4,
-            "se_ratio": 0.25,
-            "strides": 2,
-            "conv_type": 0,
-        },
-        {
-            "kernel_size": 3,
-            "num_repeat": 14,
-            "input_filters": 160,
-            "output_filters": 176,
-            "expand_ratio": 6,
-            "se_ratio": 0.25,
-            "strides": 1,
-            "conv_type": 0,
-        },
-        {
-            "kernel_size": 3,
-            "num_repeat": 18,
-            "input_filters": 176,
-            "output_filters": 304,
-            "expand_ratio": 6,
-            "se_ratio": 0.25,
-            "strides": 2,
-            "conv_type": 0,
-        },
-        {
-            "kernel_size": 3,
-            "num_repeat": 5,
-            "input_filters": 304,
-            "output_filters": 512,
-            "expand_ratio": 6,
-            "se_ratio": 0.25,
-            "strides": 1,
-            "conv_type": 0,
-        },
-    ],
-    "efficientnetv2-l": [
-        {
-            "kernel_size": 3,
-            "num_repeat": 4,
-            "input_filters": 32,
-            "output_filters": 32,
-            "expand_ratio": 1,
-            "se_ratio": 0,
-            "strides": 1,
-            "conv_type": 1,
-        },
-        {
-            "kernel_size": 3,
-            "num_repeat": 7,
-            "input_filters": 32,
-            "output_filters": 64,
-            "expand_ratio": 4,
-            "se_ratio": 0,
-            "strides": 2,
-            "conv_type": 1,
-        },
-        {
-            "kernel_size": 3,
-            "num_repeat": 7,
-            "input_filters": 64,
-            "output_filters": 96,
-            "expand_ratio": 4,
-            "se_ratio": 0,
-            "strides": 2,
-            "conv_type": 1,
-        },
-        {
-            "kernel_size": 3,
-            "num_repeat": 10,
-            "input_filters": 96,
-            "output_filters": 192,
-            "expand_ratio": 4,
-            "se_ratio": 0.25,
-            "strides": 2,
-            "conv_type": 0,
-        },
-        {
-            "kernel_size": 3,
-            "num_repeat": 19,
-            "input_filters": 192,
-            "output_filters": 224,
-            "expand_ratio": 6,
-            "se_ratio": 0.25,
-            "strides": 1,
-            "conv_type": 0,
-        },
-        {
-            "kernel_size": 3,
-            "num_repeat": 25,
-            "input_filters": 224,
-            "output_filters": 384,
-            "expand_ratio": 6,
-            "se_ratio": 0.25,
-            "strides": 2,
-            "conv_type": 0,
-        },
-        {
-            "kernel_size": 3,
-            "num_repeat": 7,
-            "input_filters": 384,
-            "output_filters": 640,
-            "expand_ratio": 6,
-            "se_ratio": 0.25,
-            "strides": 1,
-            "conv_type": 0,
-        },
-    ],
-    "efficientnetv2-b0": [
-        {
-            "kernel_size": 3,
-            "num_repeat": 1,
-            "input_filters": 32,
-            "output_filters": 16,
-            "expand_ratio": 1,
-            "se_ratio": 0,
-            "strides": 1,
-            "conv_type": 1,
-        },
-        {
-            "kernel_size": 3,
-            "num_repeat": 2,
-            "input_filters": 16,
-            "output_filters": 32,
-            "expand_ratio": 4,
-            "se_ratio": 0,
-            "strides": 2,
-            "conv_type": 1,
-        },
-        {
-            "kernel_size": 3,
-            "num_repeat": 2,
-            "input_filters": 32,
-            "output_filters": 48,
-            "expand_ratio": 4,
-            "se_ratio": 0,
-            "strides": 2,
-            "conv_type": 1,
-        },
-        {
-            "kernel_size": 3,
-            "num_repeat": 3,
-            "input_filters": 48,
-            "output_filters": 96,
-            "expand_ratio": 4,
-            "se_ratio": 0.25,
-            "strides": 2,
-            "conv_type": 0,
-        },
-        {
-            "kernel_size": 3,
-            "num_repeat": 5,
-            "input_filters": 96,
-            "output_filters": 112,
-            "expand_ratio": 6,
-            "se_ratio": 0.25,
-            "strides": 1,
-            "conv_type": 0,
-        },
-        {
-            "kernel_size": 3,
-            "num_repeat": 8,
-            "input_filters": 112,
-            "output_filters": 192,
-            "expand_ratio": 6,
-            "se_ratio": 0.25,
-            "strides": 2,
-            "conv_type": 0,
-        },
-    ],
-    "efficientnetv2-b1": [
-        {
-            "kernel_size": 3,
-            "num_repeat": 1,
-            "input_filters": 32,
-            "output_filters": 16,
-            "expand_ratio": 1,
-            "se_ratio": 0,
-            "strides": 1,
-            "conv_type": 1,
-        },
-        {
-            "kernel_size": 3,
-            "num_repeat": 2,
-            "input_filters": 16,
-            "output_filters": 32,
-            "expand_ratio": 4,
-            "se_ratio": 0,
-            "strides": 2,
-            "conv_type": 1,
-        },
-        {
-            "kernel_size": 3,
-            "num_repeat": 2,
-            "input_filters": 32,
-            "output_filters": 48,
-            "expand_ratio": 4,
-            "se_ratio": 0,
-            "strides": 2,
-            "conv_type": 1,
-        },
-        {
-            "kernel_size": 3,
-            "num_repeat": 3,
-            "input_filters": 48,
-            "output_filters": 96,
-            "expand_ratio": 4,
-            "se_ratio": 0.25,
-            "strides": 2,
-            "conv_type": 0,
-        },
-        {
-            "kernel_size": 3,
-            "num_repeat": 5,
-            "input_filters": 96,
-            "output_filters": 112,
-            "expand_ratio": 6,
-            "se_ratio": 0.25,
-            "strides": 1,
-            "conv_type": 0,
-        },
-        {
-            "kernel_size": 3,
-            "num_repeat": 8,
-            "input_filters": 112,
-            "output_filters": 192,
-            "expand_ratio": 6,
-            "se_ratio": 0.25,
-            "strides": 2,
-            "conv_type": 0,
-        },
-    ],
-    "efficientnetv2-b2": [
-        {
-            "kernel_size": 3,
-            "num_repeat": 1,
-            "input_filters": 32,
-            "output_filters": 16,
-            "expand_ratio": 1,
-            "se_ratio": 0,
-            "strides": 1,
-            "conv_type": 1,
-        },
-        {
-            "kernel_size": 3,
-            "num_repeat": 2,
-            "input_filters": 16,
-            "output_filters": 32,
-            "expand_ratio": 4,
-            "se_ratio": 0,
-            "strides": 2,
-            "conv_type": 1,
-        },
-        {
-            "kernel_size": 3,
-            "num_repeat": 2,
-            "input_filters": 32,
-            "output_filters": 48,
-            "expand_ratio": 4,
-            "se_ratio": 0,
-            "strides": 2,
-            "conv_type": 1,
-        },
-        {
-            "kernel_size": 3,
-            "num_repeat": 3,
-            "input_filters": 48,
-            "output_filters": 96,
-            "expand_ratio": 4,
-            "se_ratio": 0.25,
-            "strides": 2,
-            "conv_type": 0,
-        },
-        {
-            "kernel_size": 3,
-            "num_repeat": 5,
-            "input_filters": 96,
-            "output_filters": 112,
-            "expand_ratio": 6,
-            "se_ratio": 0.25,
-            "strides": 1,
-            "conv_type": 0,
-        },
-        {
-            "kernel_size": 3,
-            "num_repeat": 8,
-            "input_filters": 112,
-            "output_filters": 192,
-            "expand_ratio": 6,
-            "se_ratio": 0.25,
-            "strides": 2,
-            "conv_type": 0,
-        },
-    ],
-    "efficientnetv2-b3": [
-        {
-            "kernel_size": 3,
-            "num_repeat": 1,
-            "input_filters": 32,
-            "output_filters": 16,
-            "expand_ratio": 1,
-            "se_ratio": 0,
-            "strides": 1,
-            "conv_type": 1,
-        },
-        {
-            "kernel_size": 3,
-            "num_repeat": 2,
-            "input_filters": 16,
-            "output_filters": 32,
-            "expand_ratio": 4,
-            "se_ratio": 0,
-            "strides": 2,
-            "conv_type": 1,
-        },
-        {
-            "kernel_size": 3,
-            "num_repeat": 2,
-            "input_filters": 32,
-            "output_filters": 48,
-            "expand_ratio": 4,
-            "se_ratio": 0,
-            "strides": 2,
-            "conv_type": 1,
-        },
-        {
-            "kernel_size": 3,
-            "num_repeat": 3,
-            "input_filters": 48,
-            "output_filters": 96,
-            "expand_ratio": 4,
-            "se_ratio": 0.25,
-            "strides": 2,
-            "conv_type": 0,
-        },
-        {
-            "kernel_size": 3,
-            "num_repeat": 5,
-            "input_filters": 96,
-            "output_filters": 112,
-            "expand_ratio": 6,
-            "se_ratio": 0.25,
-            "strides": 1,
-            "conv_type": 0,
-        },
-        {
-            "kernel_size": 3,
-            "num_repeat": 8,
-            "input_filters": 112,
-            "output_filters": 192,
-            "expand_ratio": 6,
-            "se_ratio": 0.25,
-            "strides": 2,
-            "conv_type": 0,
-        },
-    ],
-}
-
-CONV_KERNEL_INITIALIZER = {
-    "class_name": "VarianceScaling",
-    "config": {
-        "scale": 2.0,
-        "mode": "fan_out",
-        "distribution": "truncated_normal",
-    },
-}
-
-DENSE_KERNEL_INITIALIZER = {
-    "class_name": "VarianceScaling",
-    "config": {
-        "scale": 1.0 / 3.0,
-        "mode": "fan_out",
-        "distribution": "uniform",
-    },
-}
-
-BN_AXIS = 3
-
-BASE_DOCSTRING = """Instantiates the {name} architecture.
-
-    Reference:
-    - [EfficientNetV2: Smaller Models and Faster Training](https://arxiv.org/abs/2104.00298)
-      (ICML 2021)
-
-    This function returns a Keras image classification model.
-
-    For image classification use cases, see
-    [this page for detailed examples](https://keras.io/api/applications/#usage-examples-for-image-classification-models).
-
-    For transfer learning use cases, make sure to read the
-    [guide to transfer learning & fine-tuning](https://keras.io/guides/transfer_learning/).
-
-    Args:
-        include_rescaling: bool, whether to rescale the inputs. If set
-            to `True`, inputs will be passed through a `Rescaling(1/255.0)`
-            layer.
-        include_top: bool, whether to include the fully-connected layer at the
-            top of the network. If provided, `num_classes` must be provided.
-        num_classes: optional int, number of classes to classify images into
-            (only to be specified if `include_top` is `True`).
-        weights: one of `None` (random initialization), a pretrained weight file
-            path, or a reference to pre-trained weights (e.g.
-            'imagenet/classification') (see available pre-trained weights in
-            weights.py)
-        input_shape: optional shape tuple, defaults to (None, None, 3).
-        input_tensor: optional Keras tensor (i.e. output of `layers.Input()`)
-            to use as image input for the model.
-        pooling: optional pooling mode for feature extraction when `include_top`
-            is `False`.
-            - `None` means that the output of the model will be the 4D tensor
-                output of the last convolutional block.
-            - `avg` means that global average pooling will be applied to the
-                output of the last convolutional block, and thus the output of
-                the model will be a 2D tensor.
-            - `max` means that global max pooling will be applied.
-        classifier_activation: A `str` or callable. The activation function to
-            use on the "top" layer. Ignored unless `include_top=True`. Set
-            `classifier_activation=None` to return the logits of the "top"
-            layer.
-    Returns:
-      A `keras.Model` instance.
-"""  # noqa: E501
-
-
-def round_filters(filters, width_coefficient, min_depth, depth_divisor):
-    """Round number of filters based on depth multiplier."""
-    filters *= width_coefficient
-    minimum_depth = min_depth or depth_divisor
-    new_filters = max(
-        minimum_depth,
-        int(filters + depth_divisor / 2) // depth_divisor * depth_divisor,
-    )
-    return int(new_filters)
-
-
-def round_repeats(repeats, depth_coefficient):
-    """Round number of repeats based on depth multiplier."""
-    return int(math.ceil(depth_coefficient * repeats))
-
-
-@keras.utils.register_keras_serializable(package="keras_cv.models")
-class EfficientNetV2(keras.Model):
-    """Instantiates the EfficientNetV2 architecture using given scaling
-    coefficients.
-    Args:
-        include_rescaling: bool, whether to rescale the inputs. If set
-            to `True`, inputs will be passed through a `Rescaling(1/255.0)`
-            layer.
-        include_top: bool, whether to include the fully-connected
-            layer at the top of the network.
-        width_coefficient: float, scaling coefficient for network width.
-        depth_coefficient: float, scaling coefficient for network depth.
-        default_size: integer, default input image size.
-        dropout_rate: float, dropout rate before final classifier layer.
-        drop_connect_rate: float, dropout rate at skip connections.
-        depth_divisor: integer, a unit of network width.
-        min_depth: integer, minimum number of filters.
-        bn_momentum: float. Momentum parameter for Batch Normalization layers.
-        activation: activation function.
-        blocks_args: list of dicts, parameters to construct block modules.
-        model_name: string, model name.
-        weights: one of `None` (random initialization),
-            or the path to the weights file to be loaded.
-        input_shape: optional shape tuple, defaults to (None, None, 3).
-        input_tensor: optional Keras tensor (i.e. output of `layers.Input()`)
-            to use as image input for the model.
-        pooling: optional pooling mode for feature extraction
-            when `include_top` is `False`.
-            - `None` means that the output of the model will be
-                the 4D tensor output of the
-                last convolutional layer.
-            - `avg` means that global average pooling
-                will be applied to the output of the
-                last convolutional layer, and thus
-                the output of the model will be a 2D tensor.
-            - `max` means that global max pooling will
-                be applied.
-        num_classes: optional number of classes to classify images
-            into, only to be specified if `include_top` is True.
-        classifier_activation: A `str` or callable. The activation function to
-            use on the "top" layer. Ignored unless `include_top=True`. Set
-            `classifier_activation=None` to return the logits of the "top"
-            layer.
-
-    Returns:
-      A `keras.Model` instance.
-    Raises:
-      ValueError: in case of invalid argument for `weights`,
-        or invalid input shape.
-      ValueError: if `classifier_activation` is not `"softmax"` or `None` when
-        using a pretrained top layer.
-    """
-
-    def __init__(
-        self,
-        include_rescaling,
-        include_top,
-        width_coefficient,
-        depth_coefficient,
-        default_size,
-        dropout_rate=0.2,
-        drop_connect_rate=0.2,
-        depth_divisor=8,
-        min_depth=8,
-        bn_momentum=0.9,
-        activation="swish",
-        blocks_args="default",
-        model_name="efficientnet",
-        weights=None,
-        input_shape=(None, None, 3),
-        input_tensor=None,
-        pooling=None,
-        num_classes=None,
-        classifier_activation="softmax",
-        **kwargs,
-    ):
-        if blocks_args == "default":
-            blocks_args = DEFAULT_BLOCKS_ARGS[model_name]
-
-        input_blocks_args = copy.deepcopy(blocks_args)
-
-        if weights and not tf.io.gfile.exists(weights):
-            raise ValueError(
-                "The `weights` argument should be either `None` or the path to "
-                "the weights file to be loaded. Weights file not found at "
-                f"location: {weights}"
-            )
-
-        if include_top and not num_classes:
-            raise ValueError(
-                "If `include_top` is True, you should specify `num_classes`. "
-                f"Received: num_classes={num_classes}"
-            )
-
-        if include_top and pooling:
-            raise ValueError(
-                f"`pooling` must be `None` when `include_top=True`."
-                f"Received pooling={pooling} and include_top={include_top}. "
-            )
-
-        # Determine proper input shape
-        img_input = utils.parse_model_inputs(input_shape, input_tensor)
-
-        x = img_input
-
-        if include_rescaling:
-            x = layers.Rescaling(scale=1 / 255.0)(x)
-
-        # Build stem
-        stem_filters = round_filters(
-            filters=blocks_args[0]["input_filters"],
-            width_coefficient=width_coefficient,
-            min_depth=min_depth,
-            depth_divisor=depth_divisor,
-        )
-        x = layers.Conv2D(
-            filters=stem_filters,
-            kernel_size=3,
-            strides=2,
-            kernel_initializer=CONV_KERNEL_INITIALIZER,
-            padding="same",
-            use_bias=False,
-            name="stem_conv",
-        )(x)
-        x = layers.BatchNormalization(
-            axis=BN_AXIS,
-            momentum=bn_momentum,
-            name="stem_bn",
-        )(x)
-        x = layers.Activation(activation, name="stem_activation")(x)
-
-        # Build blocks
-        blocks_args = copy.deepcopy(blocks_args)
-        b = 0
-        blocks = float(sum(args["num_repeat"] for args in blocks_args))
-
-        for i, args in enumerate(blocks_args):
-            assert args["num_repeat"] > 0
-
-            # Update block input and output filters based on depth multiplier.
-            args["input_filters"] = round_filters(
-                filters=args["input_filters"],
-                width_coefficient=width_coefficient,
-                min_depth=min_depth,
-                depth_divisor=depth_divisor,
-            )
-            args["output_filters"] = round_filters(
-                filters=args["output_filters"],
-                width_coefficient=width_coefficient,
-                min_depth=min_depth,
-                depth_divisor=depth_divisor,
-            )
-
-            repeats = round_repeats(
-                repeats=args.pop("num_repeat"),
-                depth_coefficient=depth_coefficient,
-            )
-            for j in range(repeats):
-                # The first block needs to take care of stride and filter size
-                # increase.
-                if j > 0:
-                    args["strides"] = 1
-                    args["input_filters"] = args["output_filters"]
-
-                # Determine which conv type to use:
-                block = {
-                    0: MBConvBlock(
-                        input_filters=args["input_filters"],
-                        output_filters=args["output_filters"],
-                        expand_ratio=args["expand_ratio"],
-                        kernel_size=args["kernel_size"],
-                        strides=args["strides"],
-                        se_ratio=args["se_ratio"],
-                        activation=activation,
-                        bn_momentum=bn_momentum,
-                        survival_probability=drop_connect_rate * b / blocks,
-                        name="block{}{}_".format(i + 1, chr(j + 97)),
-                    ),
-                    1: FusedMBConvBlock(
-                        input_filters=args["input_filters"],
-                        output_filters=args["output_filters"],
-                        expand_ratio=args["expand_ratio"],
-                        kernel_size=args["kernel_size"],
-                        strides=args["strides"],
-                        se_ratio=args["se_ratio"],
-                        activation=activation,
-                        bn_momentum=bn_momentum,
-                        survival_probability=drop_connect_rate * b / blocks,
-                        name="block{}{}_".format(i + 1, chr(j + 97)),
-                    ),
-                }[args["conv_type"]]
-
-                x = block(x)
-                b += 1
-
-        # Build top
-        top_filters = round_filters(
-            filters=1280,
-            width_coefficient=width_coefficient,
-            min_depth=min_depth,
-            depth_divisor=depth_divisor,
-        )
-
-        x = layers.Conv2D(
-            filters=top_filters,
-            kernel_size=1,
-            strides=1,
-            kernel_initializer=CONV_KERNEL_INITIALIZER,
-            padding="same",
-            data_format="channels_last",
-            use_bias=False,
-            name="top_conv",
-        )(x)
-        x = layers.BatchNormalization(
-            axis=BN_AXIS,
-            momentum=bn_momentum,
-            name="top_bn",
-        )(x)
-        x = layers.Activation(activation=activation, name="top_activation")(x)
-
-        if include_top:
-            x = layers.GlobalAveragePooling2D(name="avg_pool")(x)
-            if dropout_rate > 0:
-                x = layers.Dropout(dropout_rate, name="top_dropout")(x)
-            x = layers.Dense(
-                num_classes,
-                activation=classifier_activation,
-                kernel_initializer=DENSE_KERNEL_INITIALIZER,
-                bias_initializer=tf.constant_initializer(0),
-                name="predictions",
-            )(x)
-        else:
-            if pooling == "avg":
-                x = layers.GlobalAveragePooling2D(name="avg_pool")(x)
-            elif pooling == "max":
-                x = layers.GlobalMaxPooling2D(name="max_pool")(x)
-
-        inputs = img_input
-
-        # Create model.
-        super().__init__(inputs=inputs, outputs=x, **kwargs)
-
-        # All references to `self` below this line
-        if weights is not None:
-            self.load_weights(weights)
-
-        self.include_rescaling = include_rescaling
-        self.include_top = include_top
-        self.width_coefficient = width_coefficient
-        self.depth_coefficient = depth_coefficient
-        self.default_size = default_size
-        self.dropout_rate = dropout_rate
-        self.drop_connect_rate = drop_connect_rate
-        self.depth_divisor = depth_divisor
-        self.min_depth = min_depth
-        self.bn_momentum = bn_momentum
-        self.activation = activation
-        self.blocks_args = input_blocks_args
-        self.model_name = model_name
-        self.input_tensor = input_tensor
-        self.pooling = pooling
-        self.num_classes = num_classes
-        self.classifier_activation = classifier_activation
-
-    def get_config(self):
-        return {
-            "include_rescaling": self.include_rescaling,
-            "include_top": self.include_top,
-            "width_coefficient": self.width_coefficient,
-            "depth_coefficient": self.depth_coefficient,
-            "default_size": self.default_size,
-            "dropout_rate": self.dropout_rate,
-            "drop_connect_rate": self.drop_connect_rate,
-            "depth_divisor": self.depth_divisor,
-            "min_depth": self.min_depth,
-            "bn_momentum": self.bn_momentum,
-            "activation": self.activation,
-            "blocks_args": self.blocks_args,
-            "model_name": self.model_name,
-            # Remove batch dimension from `input_shape`
-            "input_shape": self.input_shape[1:],
-            "input_tensor": self.input_tensor,
-            "pooling": self.pooling,
-            "num_classes": self.num_classes,
-            "classifier_activation": self.classifier_activation,
-            "trainable": self.trainable,
-        }
-
-    @classmethod
-    def from_config(cls, config):
-        return cls(**config)
-
-
-def EfficientNetV2B0(
-    *,
-    include_rescaling,
-    include_top,
-    weights=None,
-    input_shape=(None, None, 3),
-    input_tensor=None,
-    pooling=None,
-    num_classes=None,
-    classifier_activation="softmax",
-    **kwargs,
-):
-    return EfficientNetV2(
-        include_rescaling=include_rescaling,
-        include_top=include_top,
-        width_coefficient=1.0,
-        depth_coefficient=1.0,
-        default_size=224,
-        model_name="efficientnetv2-b0",
-        weights=parse_weights(weights, include_top, "efficientnetv2b0"),
-        input_shape=input_shape,
-        input_tensor=input_tensor,
-        pooling=pooling,
-        num_classes=num_classes,
-        classifier_activation=classifier_activation,
-        **kwargs,
-    )
-
-
-def EfficientNetV2B1(
-    *,
-    include_rescaling,
-    include_top,
-    weights=None,
-    input_shape=(None, None, 3),
-    input_tensor=None,
-    pooling=None,
-    num_classes=None,
-    classifier_activation="softmax",
-    **kwargs,
-):
-    return EfficientNetV2(
-        include_rescaling=include_rescaling,
-        include_top=include_top,
-        width_coefficient=1.0,
-        depth_coefficient=1.1,
-        default_size=240,
-        model_name="efficientnetv2-b1",
-        weights=parse_weights(weights, include_top, "efficientnetv2b1"),
-        input_shape=input_shape,
-        input_tensor=input_tensor,
-        pooling=pooling,
-        num_classes=num_classes,
-        classifier_activation=classifier_activation,
-        **kwargs,
-    )
-
-
-def EfficientNetV2B2(
-    *,
-    include_rescaling,
-    include_top,
-    weights=None,
-    input_shape=(None, None, 3),
-    input_tensor=None,
-    pooling=None,
-    num_classes=None,
-    classifier_activation="softmax",
-    **kwargs,
-):
-    return EfficientNetV2(
-        include_rescaling=include_rescaling,
-        include_top=include_top,
-        width_coefficient=1.1,
-        depth_coefficient=1.2,
-        default_size=260,
-        model_name="efficientnetv2-b2",
-        weights=parse_weights(weights, include_top, "efficientnetv2b2"),
-        input_shape=input_shape,
-        input_tensor=input_tensor,
-        pooling=pooling,
-        num_classes=num_classes,
-        classifier_activation=classifier_activation,
-        **kwargs,
-    )
-
-
-def EfficientNetV2B3(
-    *,
-    include_rescaling,
-    include_top,
-    weights=None,
-    input_shape=(None, None, 3),
-    input_tensor=None,
-    pooling=None,
-    num_classes=None,
-    classifier_activation="softmax",
-    **kwargs,
-):
-    return EfficientNetV2(
-        include_rescaling=include_rescaling,
-        include_top=include_top,
-        width_coefficient=1.2,
-        depth_coefficient=1.4,
-        default_size=300,
-        model_name="efficientnetv2-b3",
-        weights=parse_weights(weights, include_top, "efficientnetv2b3"),
-        input_shape=input_shape,
-        input_tensor=input_tensor,
-        pooling=pooling,
-        num_classes=num_classes,
-        classifier_activation=classifier_activation,
-        **kwargs,
-    )
-
-
-def EfficientNetV2S(
-    *,
-    include_rescaling,
-    include_top,
-    weights=None,
-    input_shape=(None, None, 3),
-    input_tensor=None,
-    pooling=None,
-    num_classes=None,
-    classifier_activation="softmax",
-    **kwargs,
-):
-    return EfficientNetV2(
-        include_rescaling=include_rescaling,
-        include_top=include_top,
-        width_coefficient=1.0,
-        depth_coefficient=1.0,
-        default_size=384,
-        model_name="efficientnetv2-s",
-        weights=parse_weights(weights, include_top, "efficientnetv2s"),
-        input_shape=input_shape,
-        input_tensor=input_tensor,
-        pooling=pooling,
-        num_classes=num_classes,
-        classifier_activation=classifier_activation,
-        **kwargs,
-    )
-
-
-def EfficientNetV2M(
-    *,
-    include_rescaling,
-    include_top,
-    weights=None,
-    input_shape=(None, None, 3),
-    input_tensor=None,
-    pooling=None,
-    num_classes=None,
-    classifier_activation="softmax",
-    **kwargs,
-):
-    return EfficientNetV2(
-        include_rescaling=include_rescaling,
-        include_top=include_top,
-        width_coefficient=1.0,
-        depth_coefficient=1.0,
-        default_size=480,
-        model_name="efficientnetv2-m",
-        weights=parse_weights(weights, include_top, "efficientnetv2m"),
-        input_shape=input_shape,
-        input_tensor=input_tensor,
-        pooling=pooling,
-        num_classes=num_classes,
-        classifier_activation=classifier_activation,
-        **kwargs,
-    )
-
-
-def EfficientNetV2L(
-    *,
-    include_rescaling,
-    include_top,
-    weights=None,
-    input_shape=(None, None, 3),
-    input_tensor=None,
-    pooling=None,
-    num_classes=None,
-    classifier_activation="softmax",
-    **kwargs,
-):
-    return EfficientNetV2(
-        include_rescaling=include_rescaling,
-        include_top=include_top,
-        width_coefficient=1.0,
-        depth_coefficient=1.0,
-        default_size=480,
-        model_name="efficientnetv2-l",
-        weights=parse_weights(weights, include_top, "efficientnetv2l"),
-        input_shape=input_shape,
-        input_tensor=input_tensor,
-        pooling=pooling,
-        num_classes=num_classes,
-        classifier_activation=classifier_activation,
-        **kwargs,
-    )
-
-
-EfficientNetV2B0.__doc__ = BASE_DOCSTRING.format(name="EfficientNetV2B0")
-EfficientNetV2B1.__doc__ = BASE_DOCSTRING.format(name="EfficientNetV2B1")
-EfficientNetV2B2.__doc__ = BASE_DOCSTRING.format(name="EfficientNetV2B2")
-EfficientNetV2B3.__doc__ = BASE_DOCSTRING.format(name="EfficientNetV2B3")
-EfficientNetV2S.__doc__ = BASE_DOCSTRING.format(name="EfficientNetV2S")
-EfficientNetV2M.__doc__ = BASE_DOCSTRING.format(name="EfficientNetV2M")
-EfficientNetV2L.__doc__ = BASE_DOCSTRING.format(name="EfficientNetV2L")
